feat(ifql2flux): reorder expression productions so recursive case comes first
With the recursive case first there is a higher priority on extending an
expression, over reducing the base case. This is important in flux since there
are no statement terminators and therefore inherent ambiguities, such as:

a = 1.0
- 2.0

Is this one statement where a is assigned ( 1.0 - 2.0 ) or is this two
statements, with the second being - 2.0? The main flux parser will always
extend and not backtrack through the decision, so set the production priorities
so we do the same.