feat(ifql2flux): add commits on the statement production
If we encounter the division/regex ambiguity, then later encounter a parse
error due to bad syntax, it is possible for us to backtrack to the ambiguity
and successfully parse huge blocks as regexs.

(a, b, c) => {
    return a / b
}
)(

In the above example, we will backtrack and parse / b as the start of a regex.
It will end EOF, which will get reported as the location of the error. The real
error is at )(.

Adding commits on the statement production ensures we don't go backwards
through successfully parsed statements and allows for the real error location
to be reported.