feat(preserve-comments): picking up comments before and after tokens
Collecting comments before and after scanning regualar and regex tokens.
Achieving this with a clone of the _scan function, which is responsible for
calling into the the C scanner. The scanner needs to be modified so that it can
stop when it can't make progress. This allows us to pick up comments after a
token, without making progress into the next real token. Since the _scan
function is quite complicated, it's easier to achieve this with a clone, then
to look to merging afterwards.

The comments are getting stashed into two new vectors in the Token struct. Not
sure if this is the right approach, as this is my first real-world rust code.
Will fix as appropriate.

Next step is to update tests and make any necessary fixes to get tests to pass,
then refactor the two _scan functions into one.

refs #1959