feat(preserve-comments): simplified comment preservation
Now using a single linked list for preserving comments, via an optional box.
Comments are collected ahead of tokens and the list is made into a forward list
of comments. Any comments that appear at the end of the file are stored in the
EOF token.

Tests are updated with this commit. refs #1959